refactor: 패키지 구조 변경 및 불필요한 클래스 삭제
