+++ src/test/java/roomescape/controller/api/ReservationTimeControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.api;
 
 import static org.assertj.core.api.Assertions.assertThat;
